requirements.txt
# Core
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Charts & Images
matplotlib==3.8.2
pillow==10.1.0
seaborn==0.13.0

# Payment
yookassa==3.0.0

# Utils
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1